+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 ### Added
+- Added `getMemUsed()`, `getMemTotal()`, and `getMemExternEstim()` methods
 ### Fixed
+- Removed `Py_INCREF`/`Py_DECREF` on `Model` in `catchEvent`/`dropEvent` that caused memory leak for imbalanced usage
 ### Changed
 ### Removed